Split AI chat run orchestration
